requirements.txt
# Core
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2

# LangChain - UPGRADED TO 1.0+ for RAGAS compatibility
langchain==1.0.7
langchain-community==1.0.4
langchain-core==1.0.5
langgraph==0.2.62
langchain-anthropic==1.0.4
langchain-text-splitters==1.0.0

# Vector Store & Embeddings
chromadb==0.5.18
langchain-huggingface==1.0.1
# PyTorch CPU-only (pas de CUDA pour économiser ~2GB de téléchargement)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
sentence-transformers==3.2.1

# Document Loaders
pypdf==5.1.0
python-multipart==0.0.12
unstructured==0.16.7

# Evaluation - RAGAS 0.3+ with Claude/Anthropic support
ragas==0.3.9
datasets>=2.16.0
